refactor assessment operations
- split operations into several files depending on the object
