+++ client/app/bundles/course/assessment/operations/monitoring.ts
@@ -1,0 +1,8 @@
+import { MonitoringRequestData } from 'types/course/assessment/monitoring';
+
+import CourseAPI from 'api/course';
+
+export const fetchMonitoringData = async (): Promise<MonitoringRequestData> => {
+  const response = await CourseAPI.assessment.assessments.fetchMonitoringData();
+  return response.data;
+};
