+++ client/app/bundles/course/assessment/pages/AssessmentAuthenticate/index.tsx
@@ -10,6 +10,7 @@ import {
 } from 'types/course/assessment/assessments';
 import { object, string } from 'yup';
 
+import { authenticateAssessment } from 'course/assessment/operations/assessments';
 import FormTextField from 'lib/components/form/fields/TextField';
 import { setReactHookFormError } from 'lib/helpers/react-hook-form-helper';
 import { getAssessmentURL } from 'lib/helpers/url-builders';
@@ -18,7 +19,6 @@ import useTranslation from 'lib/hooks/useTranslation';
 import { formatFullDateTime } from 'lib/moment';
 import formTranslations from 'lib/translations/form';
 
-import { authenticateAssessment } from '../../operations';
 import translations from '../../translations';
 
 interface AssessmentAuthenticateProps {
